First successful approach to encapsulate Motor cursor
What I wanted to achieve was to replace this bureaucratic code proposed by
Motor:

    cursor = get_potatoes()
    while (yield cursor.fetch_next):
        potato = cursor.next_object()
        # do something with this potato

With something more user-friendly, like a blocking driver would do:

    for potato in get_potatoes():
        # do something with this potato

In my first successful attempt I came up with this:

    cursor = get_potatoes()
    for potato in (yield cursor.fetch_object()):
        # do something with this potato

Of course, the perfect approach would be:

    for potato in (yield get_potatoes()):
        # do something with this potato

The yield must be present as it is the mechanism used by Tornado to make async
io work and any call to Motor must yield a Future. It should return a Future
object, which will be handled by the @gen.coroutine decorator. When the Future
is done, the decorator will return execution to our method. At that time, the
Future's result will be passed and it should be of the collections.Iterable
type so the for..in can use it.

So, every time the for loop iterates, it will call get_potatoes() again. It must
retain a state so that the next time it's called, it returns the next document
(I'm assuming get_potatoes() is doing a find()).

The first time the method is called, it must:

* prepare a Motor cursor
* asynchronously get the first batch of documents
* immediately return a Future which will give a Iterable when done
* when the for loop calls next() on that Iterable, it should immediately return
the first document

The subsequent times the method is called it must:

* asynchronously get the next batch of documents, if necessary
* immediately return a Future which, when done, will give the same Iterable from
the first time
* when the for loop calls next() on that Iterable, it should immediately return
the next document